+++ csharp_web/Models/PanierItem.cs
@@ -14,7 +14,7 @@ namespace csharp_web.Models
         public int Id { get; set; }
         public TypeProduit Type { get; set; }
         public string Nom { get; set; }
-        public decimal Prix { get; set; }
+        public double Prix { get; set; }
         public int Quantite { get; set; }
         public string Image { get; set; }
     }
